backend/pyproject.toml
[project]
name = "stream-factory-backend"
version = "0.1.0"
description = "FastAPI service for Stream Factory"
authors = [{ name = "Stream Factory" }]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.5.2",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "openpyxl>=3.1.5",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "requests>=2.32.0",
    "apscheduler>=3.10.4",
    "openai-whisper>=20231117",
    "edge-tts>=6.1.9",
    "deep-translator>=1.11.4",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
